third_party/blink/web_tests/external/wpt/css/css-conditional/js/supports-conditionText-expected.txt

This is a testharness.js-based test.
PASS conditionText getter for @supports (color: red)
PASS conditionText getter for @supports (color : red) or ( color:blue )
PASS conditionText getter for @supports not (color: red)
FAIL conditionText getter for @supports () assert_equals: expected 1 but got 0
FAIL conditionText getter for @supports func() assert_equals: expected 1 but got 0
PASS conditionText getter for @supports ([])
PASS conditionText getter for @supports ({})
FAIL conditionText getter for @supports (()) assert_equals: expected 1 but got 0
FAIL conditionText getter for @supports (func()) assert_equals: expected 1 but got 0
PASS conditionText getter for @supports (x)
PASS conditionText getter for @supports func(x)
PASS conditionText getter for @supports ([x])
PASS conditionText getter for @supports ({x})
PASS conditionText getter for @supports ((x))
PASS conditionText getter for @supports (func(x))
Harness: the test ran to completion.
